Refactor: 상세페이지 관련 폴더 구조 정리
