+++ src/components/domains/detail/course/CourseMain/CourseMain.module.scss
