+++ src/pages/CourseDetailPage/index.tsx
@@ -1,9 +1,9 @@
-import CourseDetailInfo from "@/components/domains/DetailPage/CourseDetail/CourseDetailInfo/CourseDetailInfo";
-import CourseHeader from "@/components/domains/DetailPage/CourseDetail/Header/CourseHeader";
+import CourseHeader from "@/components/domains/detail/course/CourseHeader/CourseHeader";
+import CourseMain from "@/components/domains/detail/course/CourseMain/CourseMain";
 import DetailPageLayout from "@/layout/DetailPageLayout";
 
 const CourseDetailPage = () => {
-  return <DetailPageLayout type="course" header={<CourseHeader />} main={<CourseDetailInfo />} />;
+  return <DetailPageLayout type="course" header={<CourseHeader />} main={<CourseMain />} />;
 };
 
 export default CourseDetailPage;
